# gh-139482: Make environ.clear() efficient (O(N^2) -> O(N)).