--- a/Misc/NEWS.d/next/Library/2025-10-01-17-45-27.gh-issue-139482.yDMeEa.rst
+++ b/Misc/NEWS.d/next/Library/2025-10-01-17-45-27.gh-issue-139482.yDMeEa.rst
@@ -1,0 +1,2 @@
+Make os.environ.clear() more efficient (linear complexity instead of quadratic)
+