--- a/Misc/NEWS.d/next/Library/2025-10-01-17-45-27.gh-issue-139482.yDMeEa.rst
+++ b/Misc/NEWS.d/next/Library/2025-10-01-17-45-27.gh-issue-139482.yDMeEa.rst
@@ -1,2 +1,2 @@
-Make os.environ.clear() more efficient (linear complexity instead of quadratic)
-
+Ensure that :data:`os.environ.clear() <os.environ>`
+has linear complexity instead of quadratic complexity.