Issue with a detailed description: #139482 

---

The new implementation avoids creating a snapshot of all keys on each iteration of `os.environ.clear()`,
and instead repeatedly tries to delete keys until the environment is empty.

This closely mirrors the current behavior of environ.clear(), while being implemented more efficiently both for small and (especially) for large collections of environment variables.