@gukoff 2025-10-01
@StanFromIreland I'm happy to add tests; do you have any specific test cases in mind? 

I see that `os.environ` has the existing test coverage, while this PR is focussed on the performance, not changing the already tested behavior.