@picnixz 2025-10-01
> I see that os.environ has the existing test coverage, while this PR is focussed on the performance, not changing the already tested behavior.

Usually, we have some tests when we know that inputs will take a certain amount of time to complete (we do that for int->str conversion) but the tests may be flaky, so we only test for cases when we know that an input was meant to produce an infinite loop somewhere.

Here I don't think tests are needed. If you can provide some small benchmarks to be sure that we're not doing something worse, I think we can merge it.