@gukoff 2025-10-02
After doing a benchmark, I made a fascinaning discovery - the `unsetenv` function itself has linear complexity. On every env variable change, it scans and rewrites the whole envrironment variables block. For example, in glibc: https://codebrowser.dev/glibc/glibc/stdlib/setenv.c.html#258

In other words, we avoid quadratic time complexity in one place, but still have it in another.

---

Having said that, the current change still makes the clear() method 2x faster.

Before:

```
root@9f106f251c44:/mnt# ./python.exe bench_clearenv.py --count 100
Mean: 0.2ms ± 0.0ms
root@9f106f251c44:/mnt# ./python.exe bench_clearenv.py --count 1000
Mean: 6.5ms ± 0.2ms
root@9f106f251c44:/mnt# ./python.exe bench_clearenv.py --count 10000
Mean: 568.9ms ± 37.2ms
```

After:

```
root@43b65a4aaaa9:/mnt# ./python.exe bench_clearenv.py --count 100
Mean: 0.0ms ± 0.0ms
root@43b65a4aaaa9:/mnt# ./python.exe bench_clearenv.py --count 1000
Mean: 2.3ms ± 0.1ms
root@43b65a4aaaa9:/mnt# ./python.exe bench_clearenv.py --count 10000
Mean: 217.7ms ± 8.1ms
```

_Benchmark code: https://gist.github.com/gukoff/9a05412f193ba61798251f5ca0bc1906_

---

@picnixz @StanFromIreland  @vstinner what do you think?