@picnixz 2025-10-03
> In other words, we avoid quadratic time complexity in one place, but still have it in another.

In this case, I'm not really sure it's worth the change. There is an extension `clearenv()` that may be available but it's not part of POSIX so it's a conditional usage. I don't think having `os.clearenv()` instead of looping over `os.environ` is really a good idea (for instance `os.clearenv()` could be done but `os.environ` would still be waiting to be mutated, so it's likely better to remove the keys one by one).

Now, I wouldn't mind a 2x speed-up but I think we should indicate in the docs that `os.environ.clear` has quadratic time complexity because of the underlying glibc implementation. I don't know how it's on Windows.

Btw, please update the NEWS entry as well.