@vstinner 2025-10-08
> @picnixz @StanFromIreland @vstinner what do you think?

I'm not really impressed by benchmark numbers. I expected a x10 or x100 difference since you announced that `os.environ.clear()` has a "**quadratic** time complexity".

So I'm not fully convinced that this change is required.